+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.0.3
 - Adds PopupPostsModule
+- BetterSidebarModule sorts by most viewed guilds
 
 0.0.2
 - Adds HideVotedPostsModule